inference/requirements.txt
setuptools
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
pillow>=10.0.0
numpy>=1.26.0
tensorflow>=2.16.0
tf-keras>=2.16.0
tensorflow-hub>=0.16.0